+++ src/core/zowe/core_for_zowe_sdk/request_handler.py
@@ -134,7 +134,10 @@ class RequestHandler:
 
         Returns
         -------
-        A bytes object at the format based on Content-Type header
+        Response object at the format based on Content-Type header:
+          - `bytes` when the response is binary data
+          - object when the response is JSON text
+          - `str` when the response is plain text
         """
         contentType = self.response.headers.get("Content-Type")
         if contentType == "application/octet-stream":